refactoring, modal demo added
