+++ courses-app/src/app/features/courses/courses.component.scss
@@ -1,3 +1,7 @@
 .courses-container {
   position: relative;
 }
+
+app-info {
+  margin: 2rem;
+}
