+++ courses-app/src/app/features/courses/shared/mock-course-card-action-buttons.ts
@@ -1,0 +1,14 @@
+import {ICourseCardActionButton} from "./course-card-action-buttons-model";
+
+export const COURSE_LIST_ACTION_BUTTONS: ICourseCardActionButton[] = [
+  {
+    type: 'edit',
+    buttonText: '',
+    buttonIcon: 'fa fa-pencil',
+  },
+  {
+    type: 'delete',
+    buttonText: '',
+    buttonIcon: 'fa fa-trash',
+  },
+];
